rename allowed_exponent -> BaseDimensionExponent
